# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketing Analytics",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sites": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "List Dashboard Sites",
        "operationId": "list_dashboard_sites_api_sites_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSitesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/site-alerts": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Site Alert Settings",
        "operationId": "get_site_alert_settings_api_site_alerts_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteAlertSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Update Site Alert Settings",
        "operationId": "update_site_alert_settings_api_site_alerts_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteAlertSettingsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteAlertSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardSiteSummary": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "site_name": {
            "type": "string",
            "title": "Site Name"
          },
          "allowed_origin": {
            "type": "string",
            "title": "Allowed Origin"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "standard",
              "pro"
            ],
            "title": "Plan",
            "default": "free"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "site_name",
          "allowed_origin"
        ],
        "title": "DashboardSiteSummary"
      },
      "DashboardSitesResponse": {
        "properties": {
          "sites": {
            "items": {
              "$ref": "#/components/schemas/DashboardSiteSummary"
            },
            "type": "array",
            "title": "Sites"
          }
        },
        "type": "object",
        "required": [
          "sites"
        ],
        "title": "DashboardSitesResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SiteAlertSettingsResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "anomaly_alerts_enabled": {
            "type": "boolean",
            "title": "Anomaly Alerts Enabled"
          },
          "slack_enabled": {
            "type": "boolean",
            "title": "Slack Enabled"
          },
          "slack_webhook_url_set": {
            "type": "boolean",
            "title": "Slack Webhook Url Set"
          },
          "email_enabled": {
            "type": "boolean",
            "title": "Email Enabled"
          },
          "email_recipients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Email Recipients"
          },
          "email_delivery_configured": {
            "type": "boolean",
            "title": "Email Delivery Configured"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "anomaly_alerts_enabled",
          "slack_enabled",
          "slack_webhook_url_set",
          "email_enabled",
          "email_delivery_configured"
        ],
        "title": "SiteAlertSettingsResponse"
      },
      "SiteAlertSettingsUpdateRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "anomaly_alerts_enabled": {
            "type": "boolean",
            "title": "Anomaly Alerts Enabled",
            "default": false
          },
          "slack_enabled": {
            "type": "boolean",
            "title": "Slack Enabled",
            "default": false
          },
          "slack_webhook_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack Webhook Url"
          },
          "email_enabled": {
            "type": "boolean",
            "title": "Email Enabled",
            "default": false
          },
          "email_recipients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Email Recipients"
          }
        },
        "type": "object",
        "required": [
          "site_id"
        ],
        "title": "SiteAlertSettingsUpdateRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}